refactor: imports in dapp + shared pkg
