+++ .eslintrc
@@ -8,7 +8,7 @@
   "extends": [
     "react-app",
     "plugin:prettier/recommended",
-    // "plugin:import/typescript"
+    "plugin:import/typescript"
   ],
   "settings": {
     "import/resolver": {
