+++ packages/dapp/src/App.tsx
@@ -1,9 +1,7 @@
-import { Menus, Navbar, Widgets } from 'components';
+import { FeatureCards, Menus, Navbar, Widgets } from 'components';
+import { useInitUserTracking } from 'hooks';
 import { useEffect } from 'react';
 import { AppProvider } from './AppProvider';
-import { FeatureCards } from './components';
-import { useInitUserTracking } from './hooks';
-
 // dummy commit to check testing deployments. TO BE REMOVED
 
 export default function App() {
