+++ packages/dapp/src/AppProvider.tsx
@@ -2,9 +2,9 @@ import { ArcxAnalyticsProvider } from '@arcxmoney/analytics';
 import { CssBaseline } from '@mui/material';
 import { QueryClientProvider } from '@tanstack/react-query';
 import { BackgroundGradient } from 'components';
+import { queryClient } from 'config';
+import { I18NProvider, ThemeProvider, WalletProvider } from 'providers';
 import type { PropsWithChildren } from 'react';
-import { queryClient } from './config/queryClient';
-import { I18NProvider, ThemeProvider, WalletProvider } from './providers';
 
 export const AppProvider: React.FC<PropsWithChildren<{}>> = ({ children }) => {
   return (
