+++ packages/dapp/src/components/ChainSwitch/index.ts
@@ -1,0 +1,2 @@
+export * from './ChainSwitch';
+export * from './ChainSwitch.style';
