+++ packages/dapp/src/components/Navbar/NavbarMenuDesktop.tsx
@@ -12,7 +12,7 @@ import {
   NavbarMenuList,
   NavbarPaper,
   NavbarPopper,
-} from './Navbar.style';
+} from '.';
 interface NavbarMenuProps {
   isOpenSubMenu: boolean;
   label?: string;
