+++ packages/dapp/src/config/index.ts
@@ -1,0 +1,2 @@
+export * from './initSentry';
+export * from './queryClient';
