+++ packages/dapp/src/providers/I18nProvider.tsx
@@ -1,4 +1,4 @@
-import { defaultLang } from '@transferto/shared/src/config';
+import { defaultLang } from '@transferto/shared';
 import i18next from 'i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
 import resourcesToBackend from 'i18next-resources-to-backend';
