+++ packages/dapp/src/types/userTracking.types.ts
@@ -1,5 +1,5 @@
 import { ChainID } from '@arcxmoney/analytics';
-import { WalletAccount } from './internal.types';
+import { WalletAccount } from '.';
 
 export enum EventTrackingTool {
   ARCx,
