+++ packages/shared/src/atoms/ButtonArrowBack/ButtonArrowBack.tsx
@@ -1,7 +1,7 @@
 import ArrowBackIcon from '@mui/icons-material/ArrowBack';
 import { useTheme } from '@mui/material';
-import { getContrastAlphaColor } from '../../utils';
-import { ButtonBackArrowWrapper } from './ButtonArrowBack.style';
+import { getContrastAlphaColor } from '@utils';
+import { ButtonBackArrowWrapper } from '.';
 
 type ButtonBackArrowProps = {
   onClick?: () => void;
