+++ packages/shared/src/atoms/ButtonTransparent.style.ts
@@ -1,6 +1,6 @@
 import { alpha, styled } from '@mui/material/styles';
-import { getContrastAlphaColor } from '../utils';
-import { ButtonBase } from './ButtonBase.style';
+import { getContrastAlphaColor } from '@utils';
+import { ButtonBase } from '.';
 
 export const ButtonTransparent = styled(ButtonBase)(({ theme }) => ({
   color:
