--- packages/shared/src/atoms/ChainSwitch/index.ts
@@ -1,1 +1,0 @@
-export { ChainSwitch } from './ChainSwitch';
