+++ packages/shared/src/atoms/index.ts
@@ -3,7 +3,6 @@ export { ButtonBase } from './ButtonBase.style';
 export { ButtonPrimary } from './ButtonPrimary.style';
 export { ButtonSecondary } from './ButtonSecondary.style';
 export { ButtonTransparent } from './ButtonTransparent.style';
-export * from './ChainSwitch';
 export * from './SpotButton';
 export * from './WalletCard';
 export * from './icons';
