+++ packages/shared/src/utils/walletDigest.ts
@@ -1,4 +1,4 @@
-import type { WalletAccount } from '../types/wallet';
+import type { WalletAccount } from '@types';
 export const walletDigest = (account: WalletAccount) => {
   if (!!account.address) {
     return `${account.address.substr(0, 5)}...${account.address.substr(-4)}`;
